Make Main class no longer a static class and add support for embedded server
This means that if you start smashbattle in normal mode, an extra thread is started.
In this thread the server will be launched with a special "no_sdl" flag enabled.
This flag will make sure that it doesn't perform drawing etc.

Not everything works properly yet, I think we need to refactor everything that's static into normal classes.
Like Gameplay etc. Main was most of the work.

This is a work in progress, but so far it's pretty difficult to make sure there is no shared state between the threads.